+++ test/AssemblyUtil_server_test.py
@@ -3,7 +3,9 @@ import os
 import shutil
 import time
 import unittest
+import uuid
 from os import environ
+from pathlib import Path
 from pprint import pprint
 from configparser import ConfigParser
 
@@ -25,8 +27,7 @@ class AssemblyUtilTest(unittest.TestCase):
         config.read(config_file)
         for nameval in config.items('AssemblyUtil'):
             cls.cfg[nameval[0]] = nameval[1]
-        authServiceUrl = cls.cfg.get('auth-service-url',
-                                     'https://kbase.us/services/authorization/Sessions/Login')
+        authServiceUrl = cls.cfg['auth-service-url']
         auth_client = _KBaseAuth(authServiceUrl)
         user_id = auth_client.get_user(token)
         # WARNING: don't call any logging methods on the context object,
@@ -42,40 +43,37 @@ class AssemblyUtilTest(unittest.TestCase):
                         'authenticated': 1})
         cls.wsURL = cls.cfg['workspace-url']
         cls.wsClient = workspaceService(cls.wsURL, token=token)
+        ws = cls.wsClient.create_workspace(
+            {'workspace': 'test_AssemblyUtil_' + str(int(time.time() * 1000))})
+        cls.ws_name = ws[1]
+        cls.ws_id = ws[0]
         cls.serviceImpl = AssemblyUtil(cls.cfg)
         cls.scratch = cls.cfg['scratch']
         cls.callback_url = os.environ['SDK_CALLBACK_URL']
 
     @classmethod
     def tearDownClass(cls):
-        if hasattr(cls, 'wsName'):
-            cls.wsClient.delete_workspace({'workspace': cls.wsName})
+        if hasattr(cls, 'ws_name'):
+            cls.wsClient.delete_workspace({'workspace': cls.ws_name})
             print('Test workspace was deleted')
 
     def getWsClient(self):
-        return self.__class__.wsClient
-
-    def getWsName(self):
-        if hasattr(self.__class__, 'wsName'):
-            return self.__class__.wsName
-        suffix = int(time.time() * 1000)
-        wsName = "test_AssemblyUtil_" + str(suffix)
-        self.getWsClient().create_workspace({'workspace': wsName})
-        self.__class__.wsName = wsName
-        return wsName
+        return self.wsClient
 
     def getImpl(self):
-        return self.__class__.serviceImpl
+        return self.serviceImpl
 
     def getContext(self):
-        return self.__class__.ctx
+        return self.ctx
 
     def check_fasta_file(self, ws_obj_name, expected_file):
         assemblyUtil = self.getImpl()
 
         print('attempting download')
-        fasta = assemblyUtil.get_assembly_as_fasta(self.getContext(),
-                                                   {'ref': self.getWsName() + "/MyNewAssembly"})[0]
+        fasta = assemblyUtil.get_assembly_as_fasta(
+            self.getContext(),
+            {'ref': self.ws_name + "/" + ws_obj_name}
+        )[0]
         pprint(fasta)
         # let's compare files pointed from fasta['path'] and expected_file
         with open(expected_file, 'r') as f:
@@ -88,7 +86,13 @@ class AssemblyUtilTest(unittest.TestCase):
     def test_basic_upload_and_download(self):
         assemblyUtil = self.getImpl()
 
-        tmp_dir = self.__class__.cfg['scratch']
+        # use a temp dir for input files since when the files are downloaded to test them
+        # in _check_fasta_file they always go in the root directory, creating the potential
+        # for clobberation.
+        # Get assemblies should be updated to allow specifying a target dir
+        tmp_dir = Path(self.cfg['scratch']) / (
+            "test_basic_upload_and_download_input" + str(uuid.uuid4()))
+        os.makedirs(tmp_dir)
         file_name = "test.fna"
         shutil.copy(os.path.join("data", file_name), tmp_dir)
         fasta_path = os.path.join(tmp_dir, file_name)
@@ -96,63 +100,70 @@ class AssemblyUtilTest(unittest.TestCase):
         ws_obj_name = 'MyNewAssembly'
         result = assemblyUtil.save_assembly_from_fasta(self.getContext(),
                                                        {'file': {'path': fasta_path},
-                                                        'workspace_name': self.getWsName(),
+                                                        'workspace_name': self.ws_name,
                                                         'assembly_name': ws_obj_name,
                                                         'taxon_ref': 'ReferenceTaxons/unknown_taxon',
                                                         })
         pprint(result)
         self.check_fasta_file(ws_obj_name, fasta_path)
 
+        print('attempting upload of gzipped file using workspace id vs name')
+        # DFU can't see the file unless it's in scratch
+        shutil.copy(Path("data") / "test2.fna.gz", tmp_dir)
+        result = assemblyUtil.save_assembly_from_fasta(
+            self.getContext(),
+            {
+                'file': {'path': str(Path(tmp_dir) / "test2.fna.gz")},
+                'workspace_id': self.ws_id,
+                'assembly_name': 'MyNewAssembly_gzip',
+                'taxon_ref': 'ReferenceTaxons/unknown_taxon',
+            })
+        pprint(result)
+        self.check_fasta_file('MyNewAssembly_gzip', "data/test2.fna")
+
         print('attempting upload through shock')
-        data_file_cli = DataFileUtil(os.environ['SDK_CALLBACK_URL'])
+        data_file_cli = DataFileUtil(self.callback_url)
         shock_id = data_file_cli.file_to_shock({'file_path': fasta_path})['shock_id']
         ws_obj_name2 = 'MyNewAssembly.2'
         result2 = assemblyUtil.save_assembly_from_fasta(self.getContext(),
                                                         {'shock_id': shock_id,
-                                                         'workspace_name': self.getWsName(),
+                                                         'workspace_name': self.ws_name,
                                                          'assembly_name': ws_obj_name2
                                                          })
         pprint(result2)
         self.check_fasta_file(ws_obj_name2, fasta_path)
 
-        print('attempting upload via ftp url')
-        ftp_url = 'ftp://ftp.ensemblgenomes.org/pub/release-29/bacteria//fasta/bacteria_8_collection/acaryochloris_marina_mbic11017/dna/Acaryochloris_marina_mbic11017.GCA_000018105.1.29.dna.genome.fa.gz'
-        ws_obj_name3 = 'MyNewAssembly.3'
-        result3 = assemblyUtil.save_assembly_from_fasta(self.getContext(),
-                                                        {'ftp_url': ftp_url,
-                                                         'workspace_name': self.getWsName(),
-                                                         'assembly_name': ws_obj_name3
-                                                         })
-        pprint(result3)
         # todo: add checks here on ws object
 
-        ws_obj_name3 = 'MyNewAssembly.3'
-        result4 = assemblyUtil.export_assembly_as_fasta(self.getContext(),
-                                                        {'input_ref': self.getWsName() + '/' + ws_obj_name3})
+        result4 = assemblyUtil.export_assembly_as_fasta(
+            self.getContext(),
+            {'input_ref': self.ws_name + '/' + ws_obj_name2})
         pprint(result4)
-
 
     def test_empty_file_error_message(self):
         assemblyUtil = self.getImpl()
-        tmp_dir = self.__class__.cfg['scratch']
+        tmp_dir = self.cfg['scratch']
         file_name = "empty.FASTA"
         shutil.copy(os.path.join("data", file_name), tmp_dir)
         empty_file_path = os.path.join(tmp_dir, file_name)
         try:
             assemblyUtil.save_assembly_from_fasta(self.getContext(),
                                                     {'file': {'path': empty_file_path},
-                                                     'workspace_name': self.getWsName(),
+                                                     'workspace_name': self.ws_name,
                                                      'assembly_name': 'empty',
                                                      'taxon_ref': 'ReferenceTaxons/unknown_taxon',
                                                      "min_contig_length": 500
                                                      })
             raise Exception("Wrong error message")
-        except Exception as e:
-            self.assertEqual('No output generated by the app.'
-                         'The combination of data and input parameters '
-                         'resulted in no valid output.', str(e))
+        except ValueError as e:
+            self.assertIn(
+                'Either the original FASTA file contained no sequences or they were all '
+                + 'filtered out based on the min_contig_length parameter for file '
+                + '/kb/module/work/tmp/import_fasta_',
+                str(e))
+            self.assertIn('/empty.FASTA.filtered.fa', str(e))
 
-    """def test_legacy_contigset_download(self):
+    def test_legacy_contigset_download(self):
         ws_obj_name4 = 'LegacyContigs'
         contigset_data = {'id': 'contigset',
                           'md5': '5217cb4e8684817ba925ebe125caf54a',
@@ -172,7 +183,7 @@ class AssemblyUtilTest(unittest.TestCase):
                                        }]
                           }
 
-        obj_info = self.getWsClient().save_objects({'workspace': self.getWsName(),
+        obj_info = self.getWsClient().save_objects({'workspace': self.ws_name,
                                                     'objects': [{'type': 'KBaseGenomes.ContigSet',
                                                                  'name': ws_obj_name4,
                                                                  'data': contigset_data
@@ -181,7 +192,7 @@ class AssemblyUtilTest(unittest.TestCase):
 
         assemblyUtil = self.getImpl()
         fasta = assemblyUtil.get_assembly_as_fasta(self.getContext(),
-                                                   {'ref': self.getWsName() + '/' + obj_info[1],
+                                                   {'ref': self.ws_name + '/' + obj_info[1],
                                                     'filename': 'legacy.fa'})[0]
 
         file_name = "legacy_test.fna"
@@ -204,7 +215,7 @@ class AssemblyUtilTest(unittest.TestCase):
         ws_obj_name = 'FilteredAssembly'
         result = assemblyUtil.save_assembly_from_fasta(self.getContext(),
                                                        {'file': {'path': fasta_path},
-                                                        'workspace_name': self.getWsName(),
+                                                        'workspace_name': self.ws_name,
                                                         'assembly_name': ws_obj_name,
                                                         'min_contig_length': 9,
                                                         'external_source': 'someplace',
@@ -214,7 +225,7 @@ class AssemblyUtilTest(unittest.TestCase):
                                                         'contig_info': {'s3': {'is_circ': 0, 'description': 'somethin'}}
                                                         })
 
-        dfu = DataFileUtil(os.environ['SDK_CALLBACK_URL'])
+        dfu = DataFileUtil(self.callback_url)
         assembly = dfu.get_objects({'object_refs': [result[0]]})['data'][0]['data']
 
         self.assertEqual(len(assembly['contigs']), 1)
@@ -233,24 +244,6 @@ class AssemblyUtilTest(unittest.TestCase):
         self.assertEqual(assembly['external_source_id'], 'id')
         self.assertEqual(assembly['external_source_origination_date'], 'sunday')
 
-    def test_filtered_everything(self):
-        assemblyUtil = self.getImpl()
-
-        tmp_dir = self.__class__.cfg['scratch']
-        file_name = "legacy_test.fna"
-        shutil.copy(os.path.join("data", file_name), tmp_dir)
-        fasta_path = os.path.join(tmp_dir, file_name)
-        print('attempting upload')
-        ws_obj_name = 'FilteredAssembly'
-        with self.assertRaisesRegex(ValueError, 'There are no contigs to save, '
-                                                'thus there is no valid assembly.'):
-            assemblyUtil.save_assembly_from_fasta(self.getContext(),
-                                                       {'file': {'path': fasta_path},
-                                                        'workspace_name': self.getWsName(),
-                                                        'assembly_name': ws_obj_name,
-                                                        'min_contig_length': 500
-                                                        })
-
     def test_assembly_does_not_exist(self):
         assemblyUtil = self.getImpl()
         tmp_dir = self.__class__.cfg['scratch']
@@ -258,10 +251,13 @@ class AssemblyUtilTest(unittest.TestCase):
         fasta_path = tmp_dir + "/" + file_name
         print('attempting upload')
         ws_obj_name = 'FilteredAssembly'
-        with self.assertRaisesRegex(ValueError, 'KBase Assembly Utils tried to save an assembly'):
+        err = ("KBase Assembly Utils tried to save an assembly, but the calling "
+               + f"application specified a file \\('{tmp_dir}/not_a_real_file.fna'\\) "
+               + "that is missing. Please check the application logs for details.")
+        with self.assertRaisesRegex(ValueError, err):
             assemblyUtil.save_assembly_from_fasta(self.getContext(),
                                                        {'file': {'path': fasta_path},
-                                                        'workspace_name': self.getWsName(),
+                                                        'workspace_name': self.ws_name,
                                                         'assembly_name': ws_obj_name,
                                                         'min_contig_length': 500
-                                                        })"""
+                                                        })